Registro dei giorni e dei totali

+++ db/migrate/20190915130316_create_subscriptions.rb
@@ -5,7 +5,7 @@ class CreateSubscriptions < ActiveRecord::Migration[6.0]
       t.references :year, null: false, foreign_key: true
       t.references :group, null: true, foreign_key: true
       t.integer :anno
-      t.integer :quota
+      t.float :quota
       t.boolean :quota_versata
       t.references :day, null: false, foreign_key: true
